Use Object.assign to prevent TypeScript error
The use of the spread opertor causes issues in TypeScript 3.9.x
To avoid this issue we use Object.assign to return on object with
default values.

TypeScript error:
error TS2783: '<XYZ>' is specified more than once, so this usage will be overwritten.

See Travis build with errors:
https://travis-ci.org/github/joule-labs/joule-extension/builds/713053315

Related TypeScript issues:

https://github.com/microsoft/TypeScript/issues/39671
https://github.com/microsoft/TypeScript/issues/38535